exercise-sql-queries/requirements.txt
mysql-connector-python
Faker
pandas

